+++ src/components/Button/Chip/index.ts
@@ -1,0 +1,3 @@
+import Chip from './Chip';
+
+export default Chip;
