+++ src/components/Button/IconButton/IconButton.tsx
@@ -1,0 +1,59 @@
+/** @jsxImportSource @emotion/react */
+import { ButtonHTMLAttributes, MouseEvent, ReactNode } from 'react';
+
+import { SizeType, VariantType, getSizeStyle, getVariantStyle, iconButtonStyle } from './styles';
+
+import { css } from '@emotion/react';
+import type { CustomCSSType } from 'styles/types';
+
+type IconButtonType = ButtonHTMLAttributes<HTMLButtonElement> &
+    CustomCSSType & {
+        children: ReactNode;
+        size?: SizeType;
+        variant?: VariantType;
+        isDisabled?: boolean;
+        onClick?: (e: MouseEvent<HTMLButtonElement>) => void;
+    };
+
+/**
+ *  [UI Component] Icon Button
+ *  @param children 컴포넌트
+ *  @param type 버튼 type (button | reset | submit) [optional]
+ *  @param isDisabled 활성화여부 [optional]
+ *  @param size 버튼 크기 ('' | large | medium | small) [optional]
+ *  @param variant 버튼 스타일 ('' | outlined | contained | text) [optional]
+ *  @param onClick Click Event Handler [optional]
+ *  @returns JSX.Element
+ */
+export default function IconButton(props: IconButtonType) {
+    const {
+        children,
+        type = 'button',
+        size = '',
+        variant = '',
+        isDisabled = false,
+        onClick,
+        customCSS,
+        ...rest
+    } = props;
+
+    const handleClick = (e: React.MouseEvent<HTMLButtonElement>) => {
+        e.currentTarget.blur();
+        if (onClick) {
+            onClick(e);
+        }
+    };
+
+    return (
+        <button
+            {...rest}
+            // eslint-disable-next-line react/button-has-type
+            type={type}
+            disabled={isDisabled}
+            onClick={handleClick}
+            css={css([iconButtonStyle.iconButton, getSizeStyle(size), getVariantStyle(variant), customCSS])}
+        >
+            {children}
+        </button>
+    );
+}
